servo: Merge #5535 - Stop using int/uint in net_traits (from Ms2ger:int-net); r=jdm
Source-Repo: https://github.com/servo/servo
Source-Revision: c070ad61963963de2e0d2b35aedb1e187bbb7d4b

UltraBlame original commit: d50c64590e1a1b1b5c39e34a7c4b29efed67875b

+++ servo/components/net_traits/image/holder.rs
@@ -20,7 +20,7 @@ use url::Url;
 pub struct ImageHolder<NodeAddress> {
     url: Url,
     image: Option<Arc<Box<Image>>>,
-    cached_size: Size2D<int>,
+    cached_size: Size2D<u32>,
     local_image_cache: Arc<Mutex<LocalImageCache<NodeAddress>>>,
 }
 
@@ -55,16 +55,15 @@ impl<NodeAddress: Send + 'static> ImageHolder<NodeAddress> {
     
     
     
-    pub fn size(&self) -> Size2D<int> {
+    pub fn size(&self) -> Size2D<u32> {
         self.cached_size
     }
 
     
-    pub fn get_size(&mut self, node_address: NodeAddress) -> Option<Size2D<int>> {
+    pub fn get_size(&mut self, node_address: NodeAddress) -> Option<Size2D<u32>> {
         debug!("get_size() {}", self.url.serialize());
         self.get_image(node_address).map(|img| {
-            self.cached_size = Size2D(img.width as int,
-                                      img.height as int);
+            self.cached_size = Size2D(img.width, img.height);
             self.cached_size.clone()
         })
     }
